requirements.txt
aido-protocols-daffy==6.0.59
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==1.0.3
beautifulsoup4==4.7.1
carnivalmirror==0.6.2
cbor2==5.4.2.post1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
cloudpickle==2.0.0
coloredlogs==15.0.1
coverage==6.3.2
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
dm-tree==0.1.6
docker-pycreds==0.4.0
-e git+https://github.com/duckietown/gym-duckietown.git@7c06d1ef9671ebfefc019e9607753345f2d13407#egg=duckietown_gym_daffy
duckietown-serialization-ds1==1.0.20
duckietown-world-daffy==6.2.38
executing==0.8.3
filelock==3.6.0
fonttools==4.31.2
freetype-py==2.2.0
frozendict==2.3.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
gltflib==1.0.12
grpcio==1.43.0
gym==0.21.0
humanfriendly==10.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipython==8.2.0
jedi==0.18.1
jpeg4py==0.1.4
jsonschema==4.4.0
kiwisolver==1.4.2
lxml==4.8.0
lz4==4.0.0
Markdown==3.3.6
marshmallow==3.15.0
marshmallow-enum==1.5.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
morphys==1.0
msgpack==1.0.3
mypy-extensions==0.4.3
networkx==2.7.1
nose==1.3.7
numpy==1.22.3
opencv-python==4.5.5.64
oyaml==1.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
# pkg_resources
plotly==5.6.0
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
pybase64==1.2.1
PyContracts3==3.0.2
pycparser==2.21
PyGeometry-z6==2.1.4
pyglet==1.4.10
Pygments==2.11.2
PyOpenGL==3.1.0
pyparsing==3.0.7
pyrender==0.1.45
pyrsistent==0.18.1
python-baseconv==1.2.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
ray==1.11.0
redis==4.2.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
seaborn
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
stack-data==0.2.0
svgwrite==1.4.2
tabulate==0.8.9
tenacity==8.0.1
tensorboardX==2.5
termcolor==1.1.0
tifffile==2022.3.25
torch==1.11.0
traitlets==5.1.1
trimesh==3.10.7
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.9
validate-email==1.3
varint==1.0.2
wandb==0.12.14
wcwidth==0.2.5
webcolors==1.11.1
wrapt==1.14.0
xtermcolor==1.3
zipp==3.7.0
zuper-commons-z6==6.2.4
zuper-ipce-z6==6.1.2
zuper-nodes-z6==6.2.13
zuper-typing-z6==6.2.3
